src/node_a/requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-storage>=2.14.0
google-cloud-pubsub>=2.19.0
google-cloud-aiplatform>=1.38.0
